[tint] Cleanup failure stream redirection.
The various `Failure` classes implement a `<<` operator. So, we can just
redirect the `Failure()` object from the result instead of the `reason`
or `reason.Str()`.

The `diag::Failure` `<<` operator is updated to emit `reason.Str()` as
that's what we'll always with with a diagnostic failure.

Change-Id: I0c2b89c38c3605d2935df64d6436439c2cf4a52f
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/232175
Auto-Submit: dan sinclair <dsinclair@chromium.org>
Reviewed-by: James Price <jrprice@google.com>
Commit-Queue: dan sinclair <dsinclair@chromium.org>