[spirv-reader][ir] Handle loop header being a merge block.
A loop header block maybe the merge block for a previous instruction. In
that case, we will have the header block ID in the walk stop list
already. Instead of skipping the insert due to the existence of the key,
instead overwrite the value. In the case where the ID is in the list
already, are emitting the stop block itself, so we know its no longer a
stop block as they all get emitted after their respective instructions.

Bug: 391485718
Change-Id: Idbf8995bcb3bb6168803ae513ad4ca9b140d07e1
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/232834
Commit-Queue: dan sinclair <dsinclair@chromium.org>
Reviewed-by: James Price <jrprice@google.com>