Fix BindGroup destruction race
The destruction of slab allocated BindGroup implementations had a data
race. BindGroup::DestroyImpl() would first trigger
SlabAllocator::Deallocate() and then run the destructor. This means the
memory was returned to the SlabAllocator and could be reused to create a
new BindGroup on another thread while the destructor was running.

Fix the race by running SlabAllocator::Deallocate() after the destructor
has finished. This is accomplished by acquiring a callback from the
BindGroup before the destructor runs and invoking the callback
afterwards.

This has only impacted Android+Graphite so far since it's the only
platform where BindGroups are created/destroyed with the same
BindGroupLayout on multiple threads.

Bug: 384611797
Change-Id: I92bb52517384ddbcc742ec53a69cab9cd4078a83
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/232615
Reviewed-by: Corentin Wallez <cwallez@chromium.org>
Reviewed-by: Quyen Le <lehoangquyen@chromium.org>
Commit-Queue: Kyle Charbonneau <kylechar@google.com>