[tint][uniformity] Fix loop-in data flow edges
Edges were only being propagated when the behavior of the loop body
contained `next`. We also need to propagate from the loop-continuing
block when the body behavior was just `continue`.

Fix this by moving the logic for propagating values to loop-in nodes
next to the loop-continuing handling code, so that we do not rely on
values being propagated up the variable scope stack.

Fixed: 405600543
Change-Id: I897cabdb1ba4a21e765d669fde9240e2343a1f2b
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/232614
Commit-Queue: James Price <jrprice@google.com>
Reviewed-by: David Neto <dneto@google.com>