[tint][inspector] Unify texture builtin handling.
Many pieces of metadata about textures and samplers were gathered that
depend on a combination of the texture, the sampler and the builtin that
are used. Because handles can be passed by argument to functions, the
gathering needs to trace which handles are passed around by function
call. There were three different ways to do this and this CL unifies
them into single one. The logic of what to do with the builtin call is
also grouped in one place.

This will help support binding_array for these queries since it will
need to be handled in only one place.

The return type of SamplerTexturePairs is changed from VectorRef to
std::vector to avoid leaking the type in users of Tint.

Logic to track TextureSamplerPairs is no longer used and removed from
the resolver

Adds tint_unittest coverage for the computation of
has_texture_load_with_depth_texture and
has_depth_texture_with_non_comparison_sampler. As a side effect the
InspectorRunner is folded in InspectorBuilder so that tests don't have
to choose at the level of the fixture whether they want to use WGSL
directly or the Make/Add* methods.

Bug: 393558555, 402028624
Change-Id: I943299a58d8257a3c6fb4652113ad928da081707
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/229596
Reviewed-by: James Price <jrprice@google.com>
Commit-Queue: Corentin Wallez <cwallez@chromium.org>
Reviewed-by: dan sinclair <dsinclair@chromium.org>