[tsan] Suppress DeviceLostTest.CreateBuffer on TSAN.
When running under TSAN the `DeviceLostTest.CreateBuffer` test attempts
to allocate a buffer which is larger then the max allowed. This causes
the TSAN test to fail. Add a suppression for the test instead of
attempting to run and failing.

Bug: 407515870
Change-Id: Iedddb1e7379caae5b487fbec14747eb40908e0c4
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/233754
Reviewed-by: Corentin Wallez <cwallez@chromium.org>
Commit-Queue: dan sinclair <dsinclair@chromium.org>
Auto-Submit: dan sinclair <dsinclair@chromium.org>