Compat: unlower limits
Because we excluded Mali-T624,T628,T720 we can make the
lower limit of maxStorageBuffersPerShaderStage be 8,
maxStorageBuffersInFragmentShaders be 4,
and maxStorageTexturesInFragmentShaders be 4

I removed 2 tests because AFAICT they are hard to fix AND, they
are tested in the CTS so they'll still be tested.

The issue is the tests were testing that, in compat, when
maxStorageTexturesInFragmentStage is less than
maxStorageTexturesPerShaderStage and you try to use
more than maxStorageTexturesInFragmentStage you get an
error. This only happens in compat because in core the
2 limits are always the max of the 2.

With the change so that maxStorageTextureInFragmentStage
is 4 it's not, by default, lower than maxStorageTexturesPerShaderStage
which is also 4.

The testing infra for the dawn_unittests doesn't appear to
have any easy way to change this. To test you'd need some way
to tell dawn to return an adapter with maxStorageTexturesPerShaderStage
greater than 4. Setting that all up seems like a waste of time given
the CTS can and does test this already.

If it is important it can be done in a different CL.

Note: tested this with a CTS that expects the new limits.
Note: this will have to wait for a CTS roll or suppressions
added to expectations.

Bug: 406729256,406729255
Fixed: 406729256,406729255
Change-Id: I9c1ddbba8d1a9e65a899ef978e3b214ae6affd28
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/233398
Commit-Queue: Gregg Tavares <gman@chromium.org>
Reviewed-by: Corentin Wallez <cwallez@chromium.org>