[intrinsics] Allow setting the namespace for enums.
Add an attribute `@ns()` to the `enum` entry in the intrinsics file.
This allows setting the namespace used in the matchers for the enum, so
you can provide something like `@ns("spirv::type")` to change from
`core::` to `spirv::type::` in the generated matcher code. The default
value is `core`.

Bug:  407384500
Change-Id: I1ada830b20d03450eef6aadfd58277f9ddfe6f7c
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/233675
Reviewed-by: James Price <jrprice@google.com>
Commit-Queue: dan sinclair <dsinclair@chromium.org>