Convert --test-launcher-summary-output
Makes the necessary changes to automatically convert the
--test-launcher-summary-output= argument for tests to
--gtest_output=json: in standalone Dawn checkouts.

The former is passed in when running in a Chromium checkout in order
to output results to a known location for consumption by ResultDB.
However, it is a Chromium-specific argument handled by "base", which
is not usable in Dawn.

--gtest_output is built into gtest and functions similarly, although
its output format is different.

This change is necessary in order to have Dawn standalone builders
whose tests are compatible with ResultDB.

Bug: 385317083
Change-Id: I69e23dca4afa7a1e6da73953835b089755a7db96
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/234075
Commit-Queue: Brian Sheedy <bsheedy@google.com>
Commit-Queue: Kai Ninomiya <kainino@chromium.org>
Auto-Submit: Brian Sheedy <bsheedy@google.com>
Reviewed-by: Kai Ninomiya <kainino@chromium.org>