[tint][wgsl] Add `__constructible` to intrinsics def
This allows us to express the restriction that a type must be
constructible, which is the case for workgroupUniformLoad(). We can
then remove the special-case validation from the resolver.

Special-case the test permutation logic to expand __constructible to a
list of actual types.

Fixed: 408195607
Change-Id: Ic2b020e6fb243b0638a9225c1c2c01689bcf5940
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/234755
Commit-Queue: James Price <jrprice@google.com>
Reviewed-by: dan sinclair <dsinclair@chromium.org>