[ir][fuzz] Handle invalid workgroup sizes in fuzzer.
Then wokgroup size is checked by Dawn and Tint exe outside of the tint
backend. Add a similar check into the WGSL fuzzer so we don't fail with
a validation error.

This is not checked in the const-eval code for SubstituteOverrides
because Dawn needs the result of the const-eval to do it's own checks,
and shouldn't fail before it does so.

Bug: 408370947
Change-Id: I7ae3011c314e0d1abac068e2b34bdd95031271d5
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/235334
Auto-Submit: dan sinclair <dsinclair@chromium.org>
Reviewed-by: James Price <jrprice@google.com>
Commit-Queue: James Price <jrprice@google.com>