Break const-eval runtime semantics from the resolver.
The const-eval runtime semantics tests don't use the resolver, but did
use a couple of the helper methods provided by the builder. Remove the
includes which pull in the resolver and fix the few builder methods to
use the type and constant mangers directly.

Bug: 409319601
Change-Id: Iba7d7f3c3c511abd2dcdf1241b4895021c03da92
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/236274
Reviewed-by: James Price <jrprice@google.com>
Commit-Queue: James Price <jrprice@google.com>