Add dry-run option to cts roll
This allows you to see both the list of tests as well as the
commit message, which shows the list of pull requests that will
be included.

Note: this is useful because I tried `--test-filter "rgba8unorm"`
forgetting it's a glob. It matched nothing and generated a CL
with 0 tests which then failed. Adding this option lets you see
that your query/filter selected what you wanted. It also lets
you see what PRs are included which can be useful as sometimes
you want your tests to be on top of dawn's CTS version and not
github's top of tree version.

Bug: 409972647
Change-Id: Id28e599856f79e6b33c9a5abba857fb01214a0a5
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/236454
Reviewed-by: Brian Sheedy <bsheedy@google.com>
Commit-Queue: Gregg Tavares <gman@chromium.org>