Update SubstituteOverrideHelper to work on IR.
Chanee the SubstituteOverrideHelper to use the IR directly to get the
overrides. Instead of setting every value back into the map, only insert
the ones we require which are any which don't have an initializer.

This removes the inspector method added to get this information.

Change-Id: I0e04004d5c9c92925cc41b9b40a9f4355cfd78a1
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/236815
Commit-Queue: dan sinclair <dsinclair@chromium.org>
Reviewed-by: Peter McNeeley <petermcneeley@google.com>