Remove SubstituteOverride AST config from Dawn.
The Dawn build was using the AST configuration for SubstituteOverrides
as a way to pass the map around and then converting that to the IR
version later. Instead of passing the AST object around, pass the map
directly which is what we actually need.

Change-Id: I3c9859d529991bfb4a55cd35fb614f764c1e2834
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/237074
Commit-Queue: dan sinclair <dsinclair@chromium.org>
Reviewed-by: Corentin Wallez <cwallez@chromium.org>