[dawn][native] Support BG with BGLs that use arraySize.
This expands the single BGL entry at the API surface into one
BindingInfo per element in the BGL's internal format. This lets BG
validation and code to store element unchanged. A test is added that one
must specify all the array elements when creating a BG with a BGL that
uses arraySize.

BGL validation is updated to require that elements created for the
arrays never go past maxBindingsPerBindGroup, and a test is added.

Bug: 393558555
Change-Id: I7e91040d7659d4db26bc1b46a38aef08e707be0c
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/236554
Commit-Queue: Corentin Wallez <cwallez@chromium.org>
Reviewed-by: Brandon Jones <bajones@chromium.org>
Reviewed-by: Loko Kung <lokokung@google.com>