[dawn][native] Add Compatibility between BGLEArraySize and binding_array
This adds reflection of the array size of the WGSL bindings in
EntryPointMetadata. The compatibility check with the BGL is updated to
require the arraySize to be big enough.

The pipeline defaulting is also updated to make a BGLEArraySize when
needed.

Tests are added.

Bug: 393558555
Change-Id: I77e1ad908cd0cc48a9e54a94e2beb004f40079c8
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/236734
Commit-Queue: Corentin Wallez <cwallez@chromium.org>
Reviewed-by: Loko Kung <lokokung@google.com>
Reviewed-by: Brandon Jones <bajones@chromium.org>