Add VulkanAddWorkToEmptyResolvePass Toggle
This toggle is enabled by default on Qualcomm GPUs and adds a small
amount of work (a timestamp write) to render passes which contain no
draw commands. This works around a driver bug where the resolve targets
of the empty pass don't get resolved.

This case is covered by the CTS with
https://gpuweb.github.io/cts/standalone/?runnow=1&q=webgpu:api,operation,render_pass,resolve:render_pass_resolve:*
which was observed to fail on a Pixel 4 but with this fix now passes.

Bug: 411656647
Change-Id: If89e57df5289e18cf7314a9025d2ac985792ba8f
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/238334
Reviewed-by: Kai Ninomiya <kainino@chromium.org>
Commit-Queue: Brandon Jones <bajones@chromium.org>
Reviewed-by: Loko Kung <lokokung@google.com>