[spirv-reader][ir] Error on mixed atomic Min/Max calls.
In SPIR-V it's permitted to call an `OpAtomicU{Min,Max}` on signed
values, and `OpAtomicS{Min,Max}` on unsigned values. The min/max
operations happen as the type of the call, but there is no way to
represent this in WGSL, we'd end up doing the min/max in the type of the
arguments.

Assert that the argument type matches the call name in the SPIR-V
reader.

Bug: 406565362
Change-Id: Ic73176e120cc0562cc22728d80df7963a954234a
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/237234
Commit-Queue: dan sinclair <dsinclair@chromium.org>
Reviewed-by: James Price <jrprice@google.com>