[spirv-reader][ir] Handle UserCall atomic conversions.
When processing `UserCall` instructions for atomic conversions we may
need to fork the called functions. In order to fork, we have to collect
up all of the parameters which become atomics in that specific user call
and make a custom forked function. We store this combination of
(function, parameters converted) so we only fork the function one for
each combination of argument.

Bug: 404501988
Change-Id: I6bd318151d9e611ffae038b14c4598606f48a3ef
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/238594
Commit-Queue: dan sinclair <dsinclair@chromium.org>
Reviewed-by: James Price <jrprice@google.com>
Auto-Submit: dan sinclair <dsinclair@chromium.org>
Commit-Queue: James Price <jrprice@google.com>