Remove Node/LLVM overrides from activate-emsdk
We only use emsdk as a toolchain to build a emdawnwebgpu_pkg, samples,
and tests, so it isn't important to override the Node and LLVM binary
paths to use the ones Dawn uses. It's a bit fragile/incomplete (our LLVM
releases don't include binaries that Emscripten needs; the LLVM version
could mismatch; etc.) so it's easiest to just remove for now. We can
bring it back later if needed.

This also obsoletes a bug in activate-emsdk on Linux where PYTHON wasn't
set, so `del PYTHON` didn't work.

Bug: 371024051
Change-Id: I34a206467395de4515182fded6ec63068353f864
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/238814
Reviewed-by: Loko Kung <lokokung@google.com>
Commit-Queue: Kai Ninomiya <kainino@chromium.org>