Request contents:write permission for GitHub release workflow
The 'google' org on GitHub is configured so that the default
GITHUB_TOKEN permission for Actions is read-only. So we need to specify
which write permissions we need for the action so it can upload a
release.

Verified on my personal fork by changing the config to match 'google'.
It fails to upload a release before this fix, and succeeds after.

No-Try: true
Bug: 371024051
Change-Id: I86a772b73aaaf9984235d05215e755c777733642
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/239454
Commit-Queue: Kai Ninomiya <kainino@chromium.org>
Reviewed-by: Loko Kung <lokokung@google.com>