[tint][ir][fuzz] Re-enable generating the IR corpus
This uncomments the related action and sets its dependencies to make
sure that they are built for the host toolchain where they will be
invoked. The generation action is removed as a dependency for the
fuzzer, since it takes a long time to run and we don't want to pay
that cost on every build.

Bug: 409348527
Change-Id: I3a259ec89aa8671bfd356dd9c95c8aea491de8e8
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/239517
Reviewed-by: James Price <jrprice@google.com>
Reviewed-by: dan sinclair <dsinclair@chromium.org>
Commit-Queue: Ryan Harrison <rharrison@chromium.org>