Fix 'absl' targets for is_component_build
The upstream version of this BUILD.gn has the target as a `component`
target, which is a chromium //build specific thing, so cannot be used in
stand alone Skia.

Instead for the Dawn custom version it is a `source set` and not include
macros/defs related to component builds which were breaking Windows
specific builds.

Fixes: 409589680
Change-Id: I548d5327a48336fe836c4a5e5a678041b073fdcc
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/239794
Reviewed-by: dan sinclair <dsinclair@chromium.org>
Auto-Submit: Ryan Harrison <rharrison@chromium.org>
Commit-Queue: dan sinclair <dsinclair@chromium.org>