[tint][ir][val] Handle names for builtins without checkers
In `CheckFunctionReturnAttributesAndType` the name of the builtin is
included in the error message, which is normally fetched from the
associated `BuiltinChecker`.

For some builtins there is no checker defined, since the logic
required to validate it is more complex than what that framework
allows, which causes an ICE.

These builtins need to be handled as special cases, as-is done when
validating builtins in general.

Fixes: 414728882
Change-Id: Icd6794d7a4d415014ff311b6b496827352414a00
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/239874
Commit-Queue: Ryan Harrison <rharrison@chromium.org>
Auto-Submit: Ryan Harrison <rharrison@chromium.org>
Reviewed-by: dan sinclair <dsinclair@chromium.org>