[spirv-reader][ir] Remap samplers if needed.
When a sampler is split from an image the binding points from SPIR-V
will be duplicated. Provide the ability to pass remapping information to
the parser, or use the next largest value as to de-duplicate the values.

The tint command is extended to have a `sampler-mapping` option which
takes a space separated set of binding point pairs to remap.

Bug: 391482434
Change-Id: I13cfb6397ac45c3659e9b02eb356133e5faac043
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/241155
Reviewed-by: James Price <jrprice@google.com>
Commit-Queue: dan sinclair <dsinclair@chromium.org>