Range Analysis: Compute range of `Binary` with `Add` operator
This patch computes the range of a `Binary` with `Add` operator. To
compute the range we will first get the range of two operands, and
then compute the range of their sum. When any overflow or underflow
happen, no valid range will be returned.

Bug: 348701956
Test: tint_unittests
Change-Id: Ibe6eda96c8c654c3ffe22798472125506d595819
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/241141
Reviewed-by: James Price <jrprice@google.com>
Commit-Queue: Jiawei Shao <jiawei.shao@intel.com>