Allow WeakRef with incomplete types
Previously using WeakRef required that the full type definition was
included. This was due to the GetWeakRef() methods being a friend of the
WeakRef class so they could call the private constructor directly.

By introducing a static WeakRef<T>::Get() function which calls the
constructor and pointing the GetWeakRef calls at it we break the need
for the friend association, which in turn allows WeakRef to be used with
only a forward declaration of the type while still allowing GetWeakRef
to be called as before.

Bug: 416088623
Change-Id: I68f7d08eb1d75cff3dbce9c21459d8e6bc2b2a44
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/242276
Commit-Queue: Brandon Jones <bajones@chromium.org>
Reviewed-by: Loko Kung <lokokung@google.com>