[spirv] Cleanup image builtins in spirv.def
Cleanup some of the image builtins in the spirv.def file. Change the
`image_operands` to only accept a `u32` as it's a literal in SPIR-V. Set
all of the builtins to have named parameters to make it clearer what
values goes in each parameter.

Change-Id: I600c5d535c74c89bf9167223467c54d2683d15cf
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/242315
Commit-Queue: dan sinclair <dsinclair@chromium.org>
Reviewed-by: James Price <jrprice@google.com>