[headers] webgpu.h.diff: disable folding of unchanged sections
unified_diff lists the line numbers for each section of the diff (using
a default context of n=3 lines) which results in unchanged sections
folding like so:

  @@ -561,8 +548,19 @@

We could set n to a large value to disable folding (or use an alternate
format like ndiff that doesn't do folding), but this would make the diff
really large, and wouldn't eliminate the header of the first section:

  --- webgpu_header
  +++ dawn_header
  @@ -1,2677 +1,2724 @@

which will cause unnecessary merge conflicts between CLs that change the
header. So just remove all the line numbers to make the diff more
stable; we don't actually need this to be valid in the unified diff
format or need to be able to apply the diff to anything.

Bug: 378865841
Change-Id: I49568d782e83f409da565687df541d5d6be4be4f
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/242414
Reviewed-by: Loko Kung <lokokung@google.com>
Commit-Queue: Loko Kung <lokokung@google.com>