Implement resolve rect
'RenderPassDescriptorResolveRect' can be chained to 'RenderPassDescriptor' to
indicate that only a specified rectangular region needs to be expanded and
resolved, avoiding the need to process the entire texture.

The difference between 'RenderPassDescriptorResolveRect' and 'RenderPassDescriptorExpandResolveRect
is: 'RenderPassDescriptorResolveRect' creates an MSAA texture with dimensions no
smaller than {'width', 'height'}, which loosens the requirement of color
attachment texture size to match resolve texture size. 'RenderPassDescriptorExpandResolveRect'
creates an MSAA texture that matches the size of the resolve texture.

Bug: 402810062
Skip-Clang-Tidy-Checks: modernize-use-std-format
Change-Id: Ia5d0325944e227286be19d92f10b4230e4663121
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/238976
Commit-Queue: Xing Xu <xing.xu@intel.com>
Reviewed-by: Quyen Le <lehoangquyen@chromium.org>
Reviewed-by: Jiawei Shao <jiawei.shao@intel.com>