[tint] Add robustness for subgroup matrix builtins
They need to be predicated instead of clamped, because we cannot
guarantee that the array is large enough for a single matrix.

The stride is clamped to be at least as large as the number of rows or
columns (depending on the `col_major` parameter).

If the stride, offset, and array size are all constants, we can skip
predication.

Fixed: 403609264
Change-Id: Ic822a1e3bbeb8d28d0443433a034981bcd1185a8
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/242556
Commit-Queue: James Price <jrprice@google.com>
Reviewed-by: dan sinclair <dsinclair@chromium.org>