[emscripten] Fix fillAdapterInfoStruct__deps & externs
The deps need to be on the fillAdapterInfoStruct helper rather than the
functions that use it, I think because it's positioned like a public API
so Emscripten assumes it can't eliminate it (but still assumes it can
eliminate stringToNewUTF8). It's not an immediate code size problem
because Closure still eliminates it if it's unused, but Closure needs
the code to be valid to do that.

Tested this fixes a local release+closure build.

Also updated the externs for these; I have not verified that change (but
it's safe, it can't break anything).

No-Try: true
Bug: https://github.com/emscripten-core/emscripten/pull/24303#issuecomment-2887004258
Change-Id: Ib718f37c2dccf18374a115c9f12192685da55e46
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/242734
Commit-Queue: Loko Kung <lokokung@google.com>
Auto-Submit: Kai Ninomiya <kainino@chromium.org>
Reviewed-by: Loko Kung <lokokung@google.com>