[dawn][native][emscripten] Move/rename bindingArraySize to core struct
The handling of the new struct is equivalent to the handling of the old
struct with an extension, so this is like the extension struct is always
on. Validation is updated so that instead of the extension struct being
unsafe, bindingArraySize > 1 is unsafe. (bindingArraySize = 0 and 1 are
equivalent, and both are allowed with both array and non-array shader
bindings).

If any dependencies use non-designated initializers to create
WGPUBindGroupLayoutEntry objects (like in WireArgumentTests.cpp), this
will break compilation. Fortunately there doesn't seem to be any build
breakage in Chromium or Skia.

Fixed: 415092021
Change-Id: Ifbccfa87711e5bf5cb14242224e272a0556f7de1
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/242395
Reviewed-by: Corentin Wallez <cwallez@chromium.org>
Commit-Queue: Corentin Wallez <cwallez@chromium.org>