[WebGPU backend] Add PhysicalDeviceWGPU
Add more implementation for PhysicalDeviceWGPU.

Instance::EnumerateAdapters is not exposed in DawnProcTable.
So use wgpuInstanceRequestAdapter and pass down the
RequestAdapterOptions in BackendWGPU to create one
innerAdapter for now.

PhysicalDeviceWGPU owns this innerAdapter to make webgpu calls
and query capabilities.

Also add an extension chained struct for RequestAdapterOptions
to indicate whether to select WebGPU backend or not,
besieds the WGPUBackendType, to avoid infinite recursion in
EnumerateAdapters.

This struct can also contain an entrypoint for the proc table
in the future.

Bug: 413053623
Change-Id: Ia85aa1de579bfaee1dc198f31b1916ac7011bbe1
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/241854
Reviewed-by: Kai Ninomiya <kainino@chromium.org>
Commit-Queue: Shrek Shao <shrekshao@google.com>
Reviewed-by: Corentin Wallez <cwallez@chromium.org>