Dawn.node: Make GPUUncapturedErrorEvent extend Event
The Node NAPI provides no way to do this
https://github.com/nodejs/node-addon-api/issues/1661

To be spec compliant GPUUncapturedErrorEvent is
required to inherit from Event so it can be used
everywhere an Event can be used. The workaround
is to manually implement GPUUncapturedErrorEvent as
a JavaScript class.

Note: This class is not used in dawn.node yet but
this CL makes the CTS test
webgpu:idl,javascript:inheritance:type="GPUUncapturedErrorEvent"
pass.

A future CL will use this class.

Bug: 419128706
Depends-On: Ib56a6403bef7f677e9f5f740193770012de30ad4
Change-Id: Ia388562d8d7f02458e36b29db1a5acb8fcac03e2
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/243294
Reviewed-by: Corentin Wallez <cwallez@chromium.org>
Commit-Queue: Gregg Tavares <gman@chromium.org>