[emscripten][github] Include a "remote" emdawnwebgpu port in the release
Thanks to https://github.com/emscripten-core/emscripten/pull/24303
(which should be in the next Emscription release, 4.0.10), users can use
this new port format which just includes a pointer to a zip containing a
port file.

This change generates and releases a port file like the one used in
Emscripten, which can (1) be used instead of extracting the zip and (2)
be copied directly into Emscripten whenever we want to update its
version of `tools/ports/emdawnwebgpu.py`. It includes a full copy of the
package README as a docstring to make the port self-documenting.

Our version of Emscripten is still 4.0.8, but people can use this with
4.0.10+.

No-Try: true
Bug: 371024051
Bug: 415224043
Change-Id: I12ee39d68454d7e7dba56798864b39ade51862b0
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/243214
Auto-Submit: Kai Ninomiya <kainino@chromium.org>
Reviewed-by: Loko Kung <lokokung@google.com>
Commit-Queue: Kai Ninomiya <kainino@chromium.org>