Fix IRToProgramRoundTripTest for invalid WGSL
The `IRToProgramRoundTripTest` `RUN_TEST` method only fails if there is
an `err` string set on the result. In the case of `WGSLToIR` failing we
would just return an empty `result` and not set the error. This means,
if the WGSL if invalid, the test passes, even though it hasn't run.

This CL fixes the `result` to set the error string and then fixes up the
various tests which fail because the WGSL was invalid.

Change-Id: I39f9d4b8652a55321d7aa49aa1b5240990323de9
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/243574
Commit-Queue: dan sinclair <dsinclair@chromium.org>
Reviewed-by: James Price <jrprice@google.com>