[wgslfuzz] Validate against IR fuzzer preconditions
When wrapping an IR fuzzer with a WGSL fuzzer, we should validate
against the IR fuzzer precondition capabilities before running the
fuzzer function.

We should also not consider these validation failures to be bugs, as
they are only signalling an issue with AST->IR or substitute
overrides, both of which have their own IR fuzzers. This change is
necessary in order to validate against the preconditions as described
above.

Change-Id: I9aea47d6ed3ffe8ba032a890e0ee7d95a23e63b6
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/244158
Reviewed-by: Ryan Harrison <rharrison@chromium.org>
Reviewed-by: dan sinclair <dsinclair@chromium.org>
Commit-Queue: James Price <jrprice@google.com>