[tint] Pass array lengths via a structure
Passing the lengths buffer to all `arrayLength()` callsites and
repeatedly loading the sizes can cause compile-time explosion for some
backend compilers.

Instead, create a structure to contain all of the lengths in the entry
point and pass it down the call stack to all functions that need it.

Bug: 419595583
Change-Id: I8742306a1d94155857933ce68920bfecaa2a3fcc
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/244014
Reviewed-by: dan sinclair <dsinclair@chromium.org>
Commit-Queue: James Price <jrprice@google.com>