[spirv-reader][ir] Handle value propagation into the continue block.
A value declared in the body of a loop can propagate into the continuing
block without an `OpPhi`, in the case there is only one branch into the
continuing block. This currently breaks the propagation logic as we have
to pass the value as a block parameter and add as an operand to the
terminator instead of pushing through the control instructions.

Bug: 42250952
Change-Id: I348e3c5984907d7d9368054d2d3dd17c685b279b
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/244154
Commit-Queue: James Price <jrprice@google.com>
Reviewed-by: James Price <jrprice@google.com>
Auto-Submit: dan sinclair <dsinclair@chromium.org>