[fuzz] Remove redundant workgroup size check
This check was added in:
https://dawn-review.googlesource.com/c/dawn/+/235334

It was required to avoid the subsequent IR validation step failing and
then triggering a fuzzing issue. Now that this validation failure
causes a skip instead of a crash, we do not need the extra check in
the fuzzer.

Change-Id: I607a41d19fd533cda7bb1159ab10cb1a33fd7f79
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/244395
Reviewed-by: dan sinclair <dsinclair@chromium.org>
Commit-Queue: James Price <jrprice@google.com>