[tint] Add a SingleEntryPoint fuzzer
This transform is not run as part of the backends, so make sure we
have dedicated fuzzing coverage for it.

Make the "entry point not found" error a soft error rather than an
ICE, so that it doesn't cause a fuzzer failure.

Change-Id: Id557895e83138b43b47797c8d623013027550173
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/244159
Commit-Queue: dan sinclair <dsinclair@chromium.org>
Reviewed-by: dan sinclair <dsinclair@chromium.org>
Commit-Queue: James Price <jrprice@google.com>
Auto-Submit: James Price <jrprice@google.com>