[spirv] Rewrite MergeReturn
This change is intended to simplify the transform and make it easier
to work with, without changing any functionality.

It separates the process into two main passes:

  1. Replace all `return` instructions with `exit` instructions. Set
     the flag to signal that we are returning and record the control
     instruction that we will exit to.

  2. Conditionalize the merge blocks that we now exit to from the
     first pass. This may discover additional merge blocks that also
     need to be conditionalized as we walk back up the control flow
     stack.

The minor changes in some of the E2E tests are just a result of
iterating over instructions in a different order.

Bug: 420345727
Change-Id: I476ca8a086a99afbfd29b26f2cb01630526f5a8e
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/244474
Reviewed-by: dan sinclair <dsinclair@chromium.org>
Commit-Queue: James Price <jrprice@google.com>
Commit-Queue: dan sinclair <dsinclair@chromium.org>